Fix test breakage
