+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "prettier": "npm run prettier_vanilla -- --write",
     "prettier_check": "npm run prettier_vanilla -- --check",
-    "prettier_vanilla": "prettier '**/test/**/*.js' '**/migrations/*.js' '**/scripts/*.js' 'common/**/*.js' '*.js' 'sponsor-dapp/src/**/*.js'"
+    "prettier_vanilla": "prettier '**/test/**/*.js' '**/migrations/*.js' '**/scripts/*.js' '**/utils/*.js' 'common/**/*.js' '*.js' 'sponsor-dapp/src/**/*.js'"
   },
   "author": "UMA Team",
   "license": "MIT",
